pyproject.toml
[project]
name = "shunt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-imagekit>=6.0.0",
    "pillow>=12.0.0",
    "rich>=14.2.0",
]
